Make matplotlib optional with [plotting] extra. move emulator plotting scripts to emulators.plotting
