+++ cosmoprimo/emulators/tests/test_mlp.py
@@ -3,7 +3,8 @@ import os
 import numpy as np
 
 from cosmoprimo.fiducial import DESI
-from cosmoprimo.emulators import Emulator, EmulatedEngine, MLPEmulatorEngine, QMCSampler, Samples, FourierNormOperation, Log10Operation, PCAOperation, ChebyshevOperation, plot_residual_background, plot_residual_thermodynamics, plot_residual_harmonic, plot_residual_fourier, setup_logging
+from cosmoprimo.emulators import Emulator, EmulatedEngine, MLPEmulatorEngine, QMCSampler, Samples, FourierNormOperation, Log10Operation, PCAOperation, ChebyshevOperation, setup_logging
+from cosmoprimo.emulators.plotting import plot_residual_background, plot_residual_fourier, plot_residual_harmonic, plot_residual_thermodynamics
 
 
 samples_fn = '_tests/samples.npz'
