+++ cosmoprimo/emulators/tests/test_taylor.py
@@ -3,7 +3,8 @@ import os
 import numpy as np
 
 from cosmoprimo.fiducial import DESI
-from cosmoprimo.emulators import Emulator, EmulatedEngine, TaylorEmulatorEngine, DiffSampler, Samples, plot_residual_background, plot_residual_thermodynamics, plot_residual_harmonic, plot_residual_fourier, setup_logging
+from cosmoprimo.emulators import Emulator, EmulatedEngine, TaylorEmulatorEngine, DiffSampler, Samples, setup_logging
+from cosmoprimo.emulators.plotting import plot_residual_background, plot_residual_fourier, plot_residual_harmonic, plot_residual_thermodynamics
 
 
 samples_fn = '_tests/diff_samples.npy'
